+++ vite-project/src/components/UI/NotFound.tsx
@@ -1,4 +1,4 @@
-import { TEXT } from '../data/text';
+import { TEXT } from '../../data/text';
 
 const NotFound = () => {
   return <div>{TEXT.ERRORS.NOT_FOUND}</div>;
