+++ vite-project/src/components/users/UserList.tsx
@@ -1,5 +1,5 @@
-import { TEXT } from '../data/text';
-import { IPost } from '../data/types';
+import { TEXT } from '../../data/text';
+import { IPost } from '../../data/types';
 import UserCard from './UserCard';
 
 interface IProps {
